requirements.txt
python-magichue
odoorpc
pytz
